+++ repos/paella-ai-plugins/src/plugins/PreactButtonPlugin/PreactButtonPlugin.tsx
@@ -1,6 +1,6 @@
 import { PopUpButtonPlugin, type PopUpButtonPluginConfig } from '@asicupv/paella-core';
-import { createRoot } from 'react-dom/client';
-import { createContext, StrictMode, useContext, useRef, type ReactNode, type RefObject } from 'react';
+import { createContext, render, type ComponentChildren, type RefObject } from 'preact';
+import { useContext, useRef } from 'preact/hooks';
 import CloseIcon from "../../icons/close.svg?raw";
 
 import './PreactButtonPlugin.css';
@@ -8,17 +8,17 @@ import './PreactButtonPlugin.css';
 
 const PaellaPluginContext = createContext<PreactButtonPlugin | null>(null);
 
-export const usePaellaPlugin = () => {
+export function usePaellaPlugin<T extends PreactButtonPlugin>(): T {
   const context = useContext(PaellaPluginContext);
   if (!context) {
     throw new Error("usePaellaPlugin must be used inside PreactContainer");
   }
-  return context;
+  return context as T;
 };
 
 export const usePaellaTranslate = () => {
   const plugin = usePaellaPlugin();
-  return plugin.player.translate;
+  return plugin?.player.translate;
 };
 
 
@@ -27,10 +27,10 @@ export type PreactButtonPluginConfig = PopUpButtonPluginConfig & {
 }
 export class PreactButtonPlugin<C extends PreactButtonPluginConfig = PreactButtonPluginConfig> extends PopUpButtonPlugin<C>  {
     private _appRootElement: HTMLDivElement | null = null;
-    dialogRef: RefObject<HTMLDialogElement| null> | null = null;
+    dialogRef: RefObject<HTMLDialogElement> | null = null;
 
     async action(evt: Event, caller: HTMLElement | null = null) {
-        if (this.config.mode === "popup") {
+        if (this.config.mode === "popup" || this.config.mode === undefined) {
             return super.action(evt, caller);
         }
         // If the mode is dialog, we create the dialog element and render the AIDialog component
@@ -39,11 +39,10 @@ export class PreactButtonPlugin<C extends PreactButtonPluginConfig = PreactButto
             this._appRootElement.classList.add("PreactButtonPlugin-dialog");
     
             document.body.appendChild(this._appRootElement);
-                        
-            createRoot(this._appRootElement).render(
-                <StrictMode>
-                    <PreactDialog paellaPlugin={this} children={this.getReactNode()} />                    
-                </StrictMode>
+
+            render(
+                <PreactDialog paellaPlugin={this} children={this.getReactNode()} />,
+                this._appRootElement
             );
             // We need to wait a bit to ensure the dialog is rendered
             // Otherwise, the dialogRef will be null when we try to show it
@@ -61,17 +60,17 @@ export class PreactButtonPlugin<C extends PreactButtonPluginConfig = PreactButto
         if (this._appRootElement === null) {
             this._appRootElement = document.createElement("div");        
             this._appRootElement.classList.add("PreactButtonPlugin-popup");
-        
-            createRoot(this._appRootElement).render(
-                <StrictMode>
-                    <PreactContainer paellaPlugin={this} children={this.getReactNode()} />
-                </StrictMode>
+
+            const ReactNode = await this.getReactNode();
+            render(
+                <PreactContainer paellaPlugin={this} children={ReactNode} />,
+                this._appRootElement
             );
         }
         return this._appRootElement;
     }
 
-    getReactNode(): ReactNode {        
+    async getReactNode(): Promise<ComponentChildren> {
         return (
             <div> Hola </div>
         );
@@ -81,7 +80,7 @@ export class PreactButtonPlugin<C extends PreactButtonPluginConfig = PreactButto
 
 type PreactDialogProps = {
     paellaPlugin: PreactButtonPlugin;
-    children?: ReactNode;
+    children?: ComponentChildren;
 };
 
 const PreactDialog = ({paellaPlugin, children}: PreactDialogProps) => {
@@ -109,7 +108,7 @@ const PreactDialog = ({paellaPlugin, children}: PreactDialogProps) => {
 
 type PreactContainerProps = {
     paellaPlugin: PreactButtonPlugin;
-    children?: ReactNode;
+    children?: ComponentChildren;
 };
 
 const PreactContainer = ({paellaPlugin, children}: PreactContainerProps) => {